requirements.txt
Flask
Flask-Cors
requests
requests-oauthlib
pg8000
python-dotenv
aiohttp
flask[async]